refactor: Update get_llm function to accept optional metadata parameter
The get_llm function in ava_mosaic_ai/__init__.py has been updated to accept an optional metadata parameter. This change allows for passing additional metadata to the LLMFactory constructor when creating an instance of the LLMFactory class.

Refactor the get_llm function to accept an optional metadata parameter in order to provide more flexibility and customization when initializing the LLMFactory.